It causes too many problems
Given that python36-devel is already in ts_sal, there is no real need
to have it also on the efdwriters module, so is just deleted.
If we want to keep this definition, ts_sal class must be before this definition
otherwhise, the ts_sal module will fail because python36-devel is already
defined here.