IT-1402: merge hieradata from lsst-puppet-hiera
At present creating a branch with Puppet requires forking 3
repositories, which is quite cumbersome. Isolating Puppet code and Hiera
data is beneficial when you have 10s to 100s of Puppet developers and
want access control but we're not in that situation - so we might as
well yank it.

This commit resolves this issue by importing lsst-puppet-hiera at master
at fc88d42 and updating the hiera.yaml accordingly. We'll have a
matching commit that deletes this data from lsst-puppet-hiera.