IT-1954: set default rsyslog client log actions
In 125707a we pushed the global rsyslog configuration to the LSST
organization level which rewrote the tucson rsyslog.conf files, but we
didn't add client actions, which disabled logging to /var/log/* at
Tucson. This commit resolves the issue by pushing default configuration
at the organziation level; this configuration will be overridden on a
site-specific basis.