(profile::core::dtn) rm tql_tunning service; replace with NM dispatch
The `tql_tunning` service is being invoked on every puppet run.  This is
undesirable as it makes it appear if the puppet state on the node is
not converged.  Running interface tuning commands as a one-off is also
undesirable as parameters may be lost when the interface goes up and
down.  The NM dispatch script hooks are a more reliable option.
Additionally, interface name specific configuration generally should not
be baked into a profile class as it is inherently tied to a hosts
hardware.