(foreman role) disable ssh ProxyCommand for foreman-proxy user
When the ssh client is configured to for sssd/ipa host key
tracking using `/usr/bin/sss_ssh_knownhostsproxy` as a `ProxyCommand`,
this somehow breaks foreman-proxy REX.  Without a valid shell set in
`/etc/passwd` or via the `SHELL` env var, the ssh client errors claiming
there is a key exchange failure but never gets as far as opening a
network socket.  With a valid `SHELL`, ssh sessions are able to be
established but it appears that no stdout/stderr is piped back to the
foreman-proxy daemon.  The only identified work around is remove the
`ProxyCommand` completely for the `foreman-proxy` user.

There is some discussion of this problem on
https://github.com/theforeman/puppet-foreman_proxy/pull/742