(common) fix mit_krb5::default_cacche_name value
This key had been typoed as `mit_krb5::default_cache_nname` (cache
instead of ccache and nname instead of name) and had not taken effect.
We are relying on the default value of `FILE:/tmp/krb5cc_%{uid}` and
passing the ccache into docker containers in at least the archiver role.
Thus, we should make the default value explicit.