(spec) purge os specific facts from spec/default_fact.yml
The default fact set was copied from a live centos 7 instance.  A few of
these facts are incorrect for EL8, Debian, etc.  As well as being
largely unnecessary as part of the switch to using facterdb for test
fact sets.

Note that facterdb does not include facts from puppetlabs/stdlib and
that `#add_custom_fact` will need to be used `spec/spec_helper.rb` to
add os specific facts.