(pillan cluster) explicitly set set bond0 mac address on EL8
This is an attempt to prevent NX-OS getting upset with EL LAGs. The
theory is to prevent the bond0 mac address from changing regardless of
which sub-interface comes up first. It is unclear if this actually helps
with the strange behavior occasionally observed in the TTS but it should
not be harmful to be explicit about the LAG's mac.