(role/foreman) disable updating interfaces from puppet facts
Updating the ip / mac address for interfaces from puppet fact has caused
strange problems such as when the host grabs an incorrect IP from dhcp
and then updates the host entry... but the dhcp reservation is not
changed.  Foreman should be the source of truth and not the running
state of the host.