(rsyslog::config::actions) stop writing to /-/var/log/maillog
The leading "-" in rsyslog was an old way to disable sync-on-write.
However, this has been off by default since rsyslog v3 (c 2008).
It seems that in practice the result has been to write maillog
to "/-/var/log/maillog". Perhaps the special meaning of a leading "-"
was dropped sometime in the past decade.

Ref:
https://lsstc.slack.com/archives/C06V270ECBW/p1717180543799339
https://serverfault.com/questions/463170/what-does-filepath-action-mean-in-rsyslog-configuration
https://www.rsyslog.com/doc/compatibility/v3compatibility.html#output-file-syncing