(gha) add librarian-puppet exclusion for choria/mcollective
puppet/r10k has a dep on choria/mcollective for functionality which is
not used by lsst and this mod is not delcared in the `Puppetfile`.
choria/mcollective has dep constraints which are preventing an upgrade
of the puppet/systemd mod. librarian-puppet has the ability to ignore
mods with an `exclusion` directive. r10k does not currently support
`exclusion`. As a horrible hack, we are injecting the `exclusion`
directive only in the librarian-puppet workflow so as not to break r10k.

See: https://github.com/puppetlabs/r10k/pull/1400